Preparatory refactoring for storybook upgrade
